[all] Refactor NameDecoder to move what is relevant to BaseClass name processing in BaseClassNameHelper
