+++ Sofa/framework/Core/src/sofa/core/ObjectFactory.h
@@ -23,7 +23,7 @@
 #define SOFA_CORE_OBJECTFACTORY_H
 
 #include <sofa/core/objectmodel/BaseObject.h>
-#include <sofa/helper/NameDecoder.h>
+#include <sofa/core/objectmodel/BaseClassNameHelper.h>
 #include <numeric>
 
 namespace sofa
@@ -326,8 +326,8 @@ public:
     template<class RealObject>
     RegisterObject& add(bool defaultTemplate=false)
     {
-        std::string classname = sofa::helper::NameDecoder::getClassName<RealObject>();
-        std::string templatename = sofa::helper::NameDecoder::getTemplateName<RealObject>();
+        std::string classname = sofa::core::objectmodel::BaseClassNameHelper::getClassName<RealObject>();
+        std::string templatename = sofa::core::objectmodel::BaseClassNameHelper::getTemplateName<RealObject>();
 
         if (defaultTemplate)
             entry.defaultTemplate = templatename;
